requirements.txt
# AI NEXUS - 多模型聊天平台依赖管理
#
# 安装方法:
#   pip install -r requirements.txt
#
# 开发环境安装:
#   pip install -r requirements.txt

# ====================
# Web 框架
# ====================
flask>=3.0.0,<4.0.0
# 轻量级 Python Web 框架

flask-wtf>=1.2.1,<2.0.0
# Flask CSRF 保护和表单验证

# ====================
# HTTP 请求库
# ====================
requests>=2.31.0,<3.0.0
# Python HTTP 库，用于 SSE 流式请求

# ====================
# LLM SDK
# ====================
openai>=1.0.0,<2.0.0
# OpenAI Python SDK（兼容 DeepSeek, Moonshot 等）

google-genai>=0.3.0,<1.0.0
# Google Gemini SDK

# ====================
# 环境变量管理
# ====================
python-dotenv>=1.0.0,<2.0.0
# 从 .env 文件加载环境变量

# ====================
# 测试框架
# ====================
pytest>=7.4.0,<8.0.0
# Python 测试框架

pytest-flask>=1.2.0,<2.0.0
# Flask 测试集成

pytest-cov>=4.1.0,<5.0.0
# 代码覆盖率插件

pytest-mock>=3.11.0,<4.0.0
# Mock 和 patch 工具

# ====================
# 日志增强
# ====================
python-json-logger>=2.0.7,<3.0.0
# JSON 格式日志输出

# ====================
# 类型检查（开发）
# ====================
mypy>=1.4.0,<2.0.0
# 静态类型检查

types-requests>=2.31.0,<3.0.0
# requests 库的类型存根
